requirements.dev.txt
inmanta-core>=6
inmanta-dev-dependencies[module]
pytest-inmanta>=2
-e .[generator]
-e docs/example
